requirements.txt
# Runtime dependencies.
#
# Keep this file small: it should contain packages needed to run the GUI/CLI
# and the helper modules in src/. Test-only packages live in requirements-dev.txt.
pandas>=2.1,<3
openpyxl>=3.1,<4
PyYAML>=6,<7
PyMuPDF>=1.24,<2
PySimpleGUI>=6,<7
